refactor(canvas): split chat panel styles by owner
