+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatImageLightbox/index.css
@@ -1,0 +1,171 @@
+/* ─── Click-to-zoom images + fullscreen lightbox ───────────── */
+
+.chat-image-clickable {
+  cursor: zoom-in;
+  transition: filter 0.15s ease;
+}
+
+.chat-image-clickable:hover {
+  filter: brightness(0.94);
+}
+
+.chat-image-clickable:focus-visible {
+  outline: 2px solid #2383e2;
+  outline-offset: 2px;
+}
+
+/* Re-shelled onto ui/Modal (C1): Modal owns the portal, backdrop, ESC,
+ * dialog role, and focus trap now. `.chat-image-lightbox-card` lands on
+ * Modal's card via `className` and resets it to a transparent, full-viewport
+ * positioning shell (`position: fixed; inset: 0`) instead of Modal's default
+ * centered white card — the entrance animation (fade + slight scale) still
+ * comes from Modal's own `uiModalAppear`, so the bespoke fade-only keyframe
+ * this used to carry is gone. `.chat-image-lightbox-surface` (a plain child
+ * div, not a shell) is what actually paints the dim/blurred background and
+ * hosts the arrow-key nav + backdrop-click-to-close that Modal can't own
+ * (Modal's own backdrop sits fully behind this and is never the click
+ * target — the surface covers it edge to edge).
+ */
+.chat-image-lightbox-card {
+  /* No border-radius/box-shadow override needed: with no visible background
+   * or border of its own (transparent, borderless) and no overflow clip,
+   * Modal's default card radius/shadow render nothing visible here — only
+   * `background`/`border` actually paint, so only those are reset. */
+  position: fixed;
+  inset: 0;
+  background: transparent;
+  border: none;
+}
+
+.chat-image-lightbox-surface {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 56px 72px;
+  background: rgba(15, 15, 15, 0.82);
+  backdrop-filter: blur(4px);
+  outline: none;
+}
+
+.chat-image-lightbox-visually-hidden {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  margin: -1px;
+  padding: 0;
+  overflow: hidden;
+  clip: rect(0, 0, 0, 0);
+  white-space: nowrap;
+  border: 0;
+}
+
+.chat-image-lightbox-figure {
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 12px;
+  max-width: 100%;
+  max-height: 100%;
+}
+
+.chat-image-lightbox-figure img {
+  max-width: min(92vw, 1400px);
+  max-height: 82vh;
+  object-fit: contain;
+  border-radius: var(--radius-md);
+  box-shadow: 0 24px 64px rgba(0, 0, 0, 0.45);
+  background: #fff;
+}
+
+.chat-image-lightbox-caption {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  max-width: 92vw;
+  padding: 6px 14px;
+  border-radius: var(--radius-pill);
+  background: rgba(0, 0, 0, 0.45);
+  color: rgba(255, 255, 255, 0.92);
+  font-size: 13px;
+  line-height: 1.3;
+}
+
+.chat-image-lightbox-caption>span:first-child {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.chat-image-lightbox-counter {
+  flex-shrink: 0;
+  color: rgba(255, 255, 255, 0.6);
+  font-variant-numeric: tabular-nums;
+}
+
+.chat-image-lightbox-action {
+  position: absolute;
+  top: 18px;
+  width: 38px;
+  height: 38px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: none;
+  border-radius: 50%;
+  background: rgba(255, 255, 255, 0.12);
+  color: #fff;
+  cursor: pointer;
+  transition: background 0.15s ease;
+}
+
+.chat-image-lightbox-copy {
+  right: 64px;
+}
+
+.chat-image-lightbox-close {
+  right: 18px;
+}
+
+.chat-image-lightbox-action:hover {
+  background: rgba(255, 255, 255, 0.24);
+}
+
+.chat-image-lightbox-nav {
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  width: 44px;
+  height: 44px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: none;
+  border-radius: 50%;
+  background: rgba(255, 255, 255, 0.12);
+  color: #fff;
+  font-size: 26px;
+  line-height: 1;
+  cursor: pointer;
+  transition: background 0.15s ease;
+}
+
+.chat-image-lightbox-nav:hover {
+  background: rgba(255, 255, 255, 0.24);
+}
+
+.chat-image-lightbox-nav--prev {
+  left: 18px;
+}
+
+.chat-image-lightbox-nav--next {
+  right: 18px;
+}
+
+.chat-tool-calls--collapsed .chat-tool-calls-summary,
+.chat-tool-calls-section-header .chat-tool-calls-summary {
+  color: #73726e;
+}
