+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMentionPopup/index.css
@@ -1,0 +1,156 @@
+/* ─── @ Mention popup ────────────────────────────────────── */
+
+.chat-mention-popup {
+  position: absolute;
+  bottom: 100%;
+  left: 0;
+  right: 0;
+  margin-bottom: 4px;
+  background: #fff;
+  border: 1px solid rgba(55, 53, 47, 0.09);
+  border-radius: var(--radius-md);
+  box-shadow: 0 4px 16px rgba(15, 15, 15, 0.1), 0 0 0 1px rgba(15, 15, 15, 0.04);
+  max-height: 240px;
+  overflow-y: auto;
+  padding: 4px;
+  z-index: 20;
+}
+
+.chat-mention-group-header {
+  padding: 10px 10px 4px;
+  font-size: 10px;
+  font-weight: 500;
+  text-transform: uppercase;
+  letter-spacing: 0.6px;
+  color: rgba(55, 53, 47, 0.4);
+  pointer-events: none;
+  user-select: none;
+}
+
+.chat-mention-popup>.chat-mention-group-header:first-child {
+  padding-top: 6px;
+}
+
+.chat-mention-item {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  width: 100%;
+  padding: 6px 8px;
+  border: none;
+  border-radius: var(--radius-xs);
+  background: transparent;
+  cursor: pointer;
+  font-size: 13px;
+  color: rgb(55, 53, 47);
+  text-align: left;
+  line-height: 1.4;
+}
+
+.chat-mention-item:hover,
+.chat-mention-item--active {
+  background: rgba(55, 53, 47, 0.06);
+}
+
+.chat-mention-item-icon {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-shrink: 0;
+  width: 20px;
+  height: 20px;
+  border-radius: var(--radius-xs);
+  background: rgba(55, 53, 47, 0.06);
+  color: rgba(55, 53, 47, 0.5);
+}
+
+.chat-plugin-brand-icon,
+.chat-plugin-brand-icon img,
+.chat-plugin-brand-icon svg,
+.chat-mention-chip--plugin img,
+.chat-mention-chip--plugin svg {
+  display: block;
+  flex-shrink: 0;
+  object-fit: contain;
+}
+
+.chat-mention-item-label {
+  flex: 1;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  min-width: 0;
+}
+
+.chat-mention-item-type {
+  flex-shrink: 0;
+  font-size: 11px;
+  color: rgba(55, 53, 47, 0.35);
+  text-transform: capitalize;
+}
+
+.chat-mention-item-description {
+  flex: 2;
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  font-size: 11px;
+  color: rgba(55, 53, 47, 0.45);
+}
+
+/* Session results lead with the first user message, while workspace and date
+   remain available as compact disambiguating metadata. */
+.chat-mention-item--session .chat-mention-item-label {
+  flex: 2;
+}
+
+.chat-mention-item--session .chat-mention-item-description {
+  flex: 1;
+}
+
+/* Session previews retain structured references, but render them as compact
+   labels instead of leaking their serialized `@[...]` form and internal ids. */
+.chat-session-title {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  min-width: 0;
+  overflow: hidden;
+  white-space: nowrap;
+}
+
+/* flex-shrink: 0 — a reference chip's label is the whole point of the chip;
+   `chat-session-title-text` is what gives way when the row runs out of room.
+   Still capped at max-width (and internally ellipsized) so one long label
+   can't dominate the row on its own. */
+.chat-session-title-reference {
+  display: inline-flex;
+  align-items: center;
+  gap: 3px;
+  max-width: 132px;
+  padding: 1px 5px;
+  border: 1px solid var(--border);
+  border-radius: var(--radius-xs);
+  color: var(--text-secondary);
+  background: var(--surface-subtle);
+  flex-shrink: 0;
+  overflow: hidden;
+}
+
+.chat-session-title-reference svg {
+  flex-shrink: 0;
+}
+
+.chat-session-title-reference span {
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.chat-session-title-text {
+  min-width: 0;
+  flex-shrink: 1;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
