+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMentionPopup/index.tsx
@@ -1,12 +1,13 @@
 import { useEffect, useRef } from 'react';
-import { MENTION_GROUP_LABEL_KEY, getMentionGroupKey } from './constants';
-import type { MentionItem } from './types';
-import { MentionNodeIcon, tabMentionIconType } from './utils/mentions';
-import { roleColorSoft } from '../../utils/roleColors';
-import { useI18n } from '../../i18n';
-import { SessionTitle } from './SessionTitle';
-import { sessionTitleText } from './utils/sessionTitle';
-import { pluginMentionIconMarkup } from './utils/pluginMentionIcons';
+import './index.css';
+import { MENTION_GROUP_LABEL_KEY, getMentionGroupKey } from '../constants';
+import type { MentionItem } from '../types';
+import { MentionNodeIcon, tabMentionIconType } from '../utils/mentions';
+import { roleColorSoft } from '../../../utils/roleColors';
+import { useI18n } from '../../../i18n';
+import { SessionTitle } from '../SessionTitle';
+import { sessionTitleText } from '../utils/sessionTitle';
+import { pluginMentionIconMarkup } from '../utils/pluginMentionIcons';
 
 interface ChatMentionPopupProps {
   mentionItems: MentionItem[];
