+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/index.tsx
@@ -4,7 +4,6 @@ import { useRightDock, useRightDockState } from '../../dock/RightDock/context';
 import { isDockContentTabVisible, toggleFullPageDockContentTabs } from '../../dock/RightDock/dock-content-tabs';
 import type { SettingsSection } from '../../settings/Settings';
 import '../ChatPage.css';
-import '../ChatPanel.css';
 import type { UnifiedSession } from '../ChatSessionsRail';
 import { ChatView } from '../ChatView';
 import { SessionBackBar, type SessionBackEntry } from '../SessionBackBar';
