+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatView/index.tsx
@@ -6,6 +6,7 @@ import {
   type ReactNode,
   type RefObject,
 } from 'react';
+import './index.css';
 import type {
   AgentChatMessage,
   AgentScope,
